ai/requirements.txt
flask
tensorflow
pillow
numpy
requests